Inclusão de novos extratores de dados para o Facebook Ad, Google Search Console, Google Play Store e SFTP

+++ extractor/shell/bigquery.sh
@@ -242,9 +242,11 @@ full_load()
 # Realiza a limpeza dos arquivos temporários.
 clean_up()
 {
-    echo "Removing temporary folter ${QUEUE_PATH}"
-    cd ${GLOVE_TEMP}
-    rm -rf ${QUEUE_PATH}
+    if [ ${MODULE} != "file" ]; then
+        echo "Removing temporary folder ${QUEUE_PATH}"
+        cd ${GLOVE_TEMP}
+        rm -rf ${QUEUE_PATH}
+    fi
 }
 
 # Identifica a ocorrência de erros e interrompe processo.
